Standardize Tuner subclassing
