+++ kerastuner/engine/hypermodel.py
@@ -17,6 +17,16 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+import contextlib
+import gc
+import traceback
+
+import numpy as np
+from tensorflow import keras
+
+from ..abstractions import display
+from .. import config as config_module
+
 
 class HyperModel(object):
 
@@ -32,3 +42,112 @@ class DefaultHyperModel(HyperModel):
     def __init__(self, build, name=None, tunable=True):
         super(DefaultHyperModel, self).__init__(name=name)
         self.build = build
+
+
+class KerasModelWrapper(HyperModel):
+    """Builds and compiles a Keras Model with optional compile overrides."""
+
+    def __init__(self,
+                 hypermodel,
+                 max_model_size=None,
+                 optimizer=None,
+                 loss=None,
+                 metrics=None,
+                 distribution_strategy=None):
+        self.hypermodel = get_hypermodel(hypermodel)
+        self.max_model_size = max_model_size
+        self.optimizer = optimizer
+        self.loss = loss
+        self.metrics = metrics
+        self.distribution_strategy = distribution_strategy
+
+        self._max_fail_streak = 5
+
+    def build(self, hp):
+        for i in range(self._max_fail_streak + 1):
+            # clean-up TF graph from previously stored (defunct) graph
+            keras.backend.clear_session()
+            gc.collect()
+
+            # Build a model, allowing max_fail_streak failed attempts.
+            try:
+                with maybe_distribute(self.distribution_strategy):
+                    model = self.hypermodel.build(hp)
+            except:
+                if config_module.DEBUG:
+                    traceback.print_exc()
+
+                display.warning('Invalid model %s/%s' %
+                                (i, self._max_fail_streak))
+
+                if i == self._max_fail_streak:
+                    raise RuntimeError(
+                        'Too many failed attempts to build model.')
+                continue
+
+            # Stop if `build()` does not return a valid model.
+            if not isinstance(model, keras.models.Model):
+                raise RuntimeError(
+                    'Model-building function did not return '
+                    'a valid Keras Model instance, found {}'.format(model))
+
+            # Check model size.
+            size = maybe_compute_model_size(model)
+            if self.max_model_size and size > self.max_model_size:
+                display.warning(
+                    'Oversized model: %s parameters -- skipping' % (size))
+                if i == self._max_fail_streak:
+                    raise RuntimeError(
+                        'Too many consecutive oversized models.')
+                continue
+            break
+
+        return self._compile_model(model)
+
+    def _compile_model(self, model):
+        with maybe_distribute(self.distribution_strategy):
+            if self.optimizer or self.loss or self.metrics:
+                compile_kwargs = {
+                    'optimizer': model.optimizer,
+                    'loss': model.loss,
+                    'metrics': model.metrics,
+                }
+                if self.loss:
+                    compile_kwargs['loss'] = self.loss
+                if self.optimizer:
+                    compile_kwargs['optimizer'] = self.optimizer
+                if self.metrics:
+                    compile_kwargs['metrics'] = self.metrics
+                model.compile(**compile_kwargs)
+            return model
+
+
+def maybe_compute_model_size(model):
+    """Compute the size of a given model, if it has been built."""
+    if model.built:
+        params = [keras.backend.count_params(p) for p in model.trainable_weights]
+        return int(np.sum(params))
+    return 0
+
+
+@contextlib.contextmanager
+def maybe_distribute(distribution_strategy):
+    """Distributes if distribution_strategy is set."""
+    if distribution_strategy is None:
+        yield
+    else:
+        with distribution_strategy.scope():
+            yield
+
+
+def get_hypermodel(hypermodel):
+    """Gets a HyperModel from a HyperModel or callable."""
+    if isinstance(hypermodel, HyperModel):
+        return hypermodel
+    else:
+        if not callable(hypermodel):
+            raise ValueError(
+                'The `hypermodel` argument should be either '
+                'a callable with signature `build(hp)` returning a model, '
+                'or an instance of `HyperModel`.')
+        return DefaultHyperModel(hypermodel)
